+++ CHANGELOG.md
@@ -19,6 +19,10 @@ interface BaseFactory<T : BaseFactory<T>>
 class HomeFactory(...) : BaseFactory<HomeFactory>
 ```
 
+### Enhancements
+
+- Support top-level FIR gen (contribution hints, function inject, etc) in Kotlin/JS on `2.3.21`+ and `2.4.0-Beta2`+.
+
 ### Changes
 
 - Update shaded `androidx.tracing` to 2.0.0-alpha04.